[Glibc] Declare `SwiftGlibc` as `no_undeclared_includes`
This fixes a circular dependency between libstdc++ and glibc modules.

Glibc contains multiple include directives like `#include <stdlib.h>`. The problem is that `stdlib.h` is not only a glibc header, it is also a libstdc++ header that forwards to the glibc header. This causes Clang to assume that glibc depends on libstdc++.