Disallow directly reflecting closures
The Swift compiler uses various calling conventions for closures depending
on context.  This context disappears when you try to use Mirror to reflect
a closure-valued property, which can make the closure un-callable.

There doesn't seem to be any good solution to this, unless we
dictate that a closure's ABI is determined by it's type, which
we've so far been unwilling to guarantee.

So instead, this PR erases closure values during reflection,
making it impossible to directly reflect closures.
Note that the field still appears in the reflection list,
but the value is replaced with `()` (the empty tuple, aka Void).

```
struct A { var prop: () -> Int }
let a = A { return 7 }
let m = Mirror(reflecting: a)
let c = m.children.first!.value
let p = c as? () -> Int
// Before this PR, the cast above would succeed,
// but the call below could crash.
p()
```

Note however, that it _does_ work to put a closure in an `Any` or
other container, reflect that container, and then access the closure.
(In this case, the compiler has full visibility into all direct
accesses to the closure.)

```
struct B { var prop: Any }
let b = B(prop: { return 7 } as Any)
let m = Mirror(reflecting: b)
let c = m.children.first!.value
let p = c as? () -> Int // This always worked
p() // This always succeeded
```

Resolves rdar://59791962, SR-12268, SR-14362, #54696