Set back deploy minimum OS versions in cmake
This is an attempt to fix https://github.com/apple/swift/issues/58271
which appears to be a bug in however Apple internally builds the
toolchains shipped with Xcode, as it doesn't reproduce locally nor in
the toolchains on swift.org. My assumption here is that it's more likely
that these values are correctly taken into account if they live directly
in the relevant cmake configuration vs in the build scripts.