Fix update-checkout error reporting
In the event of a failure, update-checkout wouldn't always fail very
gracefully. As a result the actual failure wasn't always printed, which
makes it really difficult to figure out what's going on.

In my case, I'm jumping between 5.6 and main branches, so
the `swift-cmark-gfm` is causing update-checkout to hiccup since it's
not in the `main` configuration file scheme for `main`, but is in the
5.6 repository. You wouldn't know that though, since it would complain
about `repo_path` not being in the error message before crashing.

I've gone ahead and done some cleanups, reporting what the problem is at
the point of failure beyond the "key error 'swift-cmark-cfg'. Modernized
the error reporter a bit, so it should be a smidge easier to read.