[move-keyword] Make sure we error correctly on f(_move y, &y) and f(&y, _move y)
This was actually crashing on the inout use since I at some point changed the
isReinit code to use a more general routine to determine that which was
incorrect. Instead, this is actually trying to recognize reinits that we know
how to change into a non-reinit form (e.x.: store [assign] -> store [init]).
Since the inout use was not handled by the convert reinit to init code, we
crashed.

Now we get the correct behavior since the access to y always occurs /before/ the
formal evaluation of the inout.