[SILInliner] Handled null access storage base.
During inlining, lexical-ness of arguments is preserved.  In the case of
address arguments that come from alloc_stacks, the alloc_stacks are
marked lexical.  The base of the storage which is passed to the address
is obtained via AccessStorageWithBase.  In some circumstances, a base is
not found.  Handled that case by changing a dyn_cast of the base to a
dyn_cast_or_null.

rdar://99087653